+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/state/archive/StalledCheckpointUploadHandler.java
@@ -23,6 +23,7 @@ package com.walmartlabs.concord.server.process.state.archive;
 import com.walmartlabs.concord.db.AbstractDao;
 import com.walmartlabs.concord.server.PeriodicTask;
 import com.walmartlabs.concord.server.cfg.ProcessCheckpointArchiveConfiguration;
+import com.walmartlabs.concord.server.task.ScheduledTask;
 import org.jooq.Configuration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,12 +32,13 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 import java.sql.Timestamp;
+import java.util.concurrent.TimeUnit;
 
 import static com.walmartlabs.concord.server.jooq.Tables.PROCESS_CHECKPOINT_ARCHIVE;
 
-@Named
+@Named("stalled-checkpoint-uploader")
 @Singleton
-public class StalledCheckpointUploadHandler extends PeriodicTask {
+public class StalledCheckpointUploadHandler implements ScheduledTask {
 
     private static final Logger log = LoggerFactory.getLogger(StalledCheckpointUploadHandler.class);
 
@@ -45,13 +47,17 @@ public class StalledCheckpointUploadHandler extends PeriodicTask {
 
     @Inject
     public StalledCheckpointUploadHandler(ProcessCheckpointArchiveConfiguration cfg, CleanupDao cleanupDao) {
-        super(cfg.isEnabled() ? 60000 : 0, 30000);
         this.cfg = cfg;
         this.cleanupDao = cleanupDao;
     }
 
     @Override
-    protected void performTask() {
+    public long getIntervalInSec() {
+        return cfg.isEnabled() ? TimeUnit.MINUTES.toSeconds(1) : 0;
+    }
+
+    @Override
+    public void performTask() {
         Timestamp cutoff = new Timestamp(System.currentTimeMillis() - cfg.getStalledAge());
         cleanupDao.process(cutoff);
     }